baserole needs a group
